+++ editors/neovim/ftdetect/mcc.lua
@@ -1,0 +1,7 @@
+-- Filetype detection for mcc source files (.mc) and interface stubs (.mci).
+vim.filetype.add({
+  extension = {
+    mc = 'mcc',
+    mci = 'mcc',
+  },
+})